extra features added
1. Background: Set the overall page to use a background image of: silverware.jpg
The image should repeat in both directions across the page and should not move when the page is scrolled.
2. Favicon: Set the page to have a "favorites icon" ("favicon"). Use: pie-icon.gif
Favicons are explained in Chapter 3. The icon may not work in Internet Explorer; you may ignore this.
3. Pie bullet: Set all bulleted lists of items on the page to use an image for their bullet icon rather than the
normal black circle. Use the following image: pie-icon.gif
4. Wide headings: Place 0.25em horizontal spacing between neighboring letters in all headings on the page.
5. Tight heading background: Make it so that the gray background behind the headings on the page is only
behind the text itself, not stretched across the entire width of the page. (Looks nice with extra feature #1.)
6. Other: Do you have an extra feature you'd like to add to your page that isn't listed here? Ask us on the
course message board and we'll let you know if it is okay to substitute for one of the above.